feat: include generic price feeds in responses
