+++ Cargo.toml
@@ -27,9 +27,10 @@ ed25519-dalek = { version = "2.1" }
 frost-ed25519 = { version = "2.1", features = ["serialization"] }
 futures = "0.3"
 hex = "0.4.3"
+itertools = "0.14"
 kupon = { git = "https://github.com/SundaeSwap-finance/kupon", rev = "ccc0b37" }
-minicbor = { version = "0.25", features = ["derive", "std"] }
-minicbor-io = { version = "0.20", features = ["async-io"] }
+minicbor = { version = "0.26", features = ["derive", "std"] }
+minicbor-io = { version = "0.21", features = ["async-io"] }
 num-bigint = "0.4"
 num-integer = "0.1"
 num-rational = "0.4"
@@ -37,8 +38,8 @@ num-traits = "0.2"
 opentelemetry = "0.27"
 opentelemetry-otlp = { version = "0.27", features = ["gzip-tonic", "tls-webpki-roots"] }
 opentelemetry_sdk = { version = "0.27", features = ["rt-tokio"] }
-pallas-crypto = "0.32"
-pallas-primitives = "0.32"
+pallas-crypto = { git = "https://github.com/txpipe/pallas", rev = "c3aad16" }
+pallas-primitives = { git = "https://github.com/txpipe/pallas", rev = "c3aad16" }
 rand = "0.8.5"
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
 rust_decimal = "1.34.3"
