+++ src/signature_aggregator.rs
@@ -27,7 +27,7 @@ use crate::{
     config::OracleConfig,
     network::{Network, NodeId},
     price_feed::{
-        cbor_encode_in_list, IntervalBoundType, PriceData, Signed, SignedEntries, SyntheticEntry,
+        cbor_encode_in_list, GenericPriceFeed, IntervalBoundType, PriceData, Signed, SignedEntries,
         SyntheticPriceFeed,
     },
     raft::{RaftClient, RaftLeader},
@@ -36,7 +36,7 @@ use crate::{
 pub struct SignatureAggregator {
     id: NodeId,
     implementation: SignatureAggregatorImplementation,
-    synthetics: Vec<String>,
+    feeds: Vec<String>,
     raft_client: RaftClient,
     signed_entries_source: mpsc::Receiver<(NodeId, SignedEntries)>,
     payload_sink: watch::Sender<Payload>,
@@ -93,12 +93,18 @@ impl SignatureAggregator {
             mpsc::Sender<(NodeId, SignedEntries)>,
         ) -> Result<SignatureAggregatorImplementation>,
     {
-        let synthetics = config.synthetics.iter().map(|s| s.name.clone()).collect();
+        let feeds = config
+            .synthetics
+            .iter()
+            .map(|s| s.name.clone())
+            .chain(config.feeds.all_feeds().iter().map(|f| f.name()))
+            .collect();
         let (signed_entries_sink, signed_entries_source) = mpsc::channel(10);
         let (payload_sink, mut payload_source) = watch::channel(Payload {
             publisher: config.id.clone(),
             timestamp: SystemTime::UNIX_EPOCH,
             synthetics: vec![],
+            generics: vec![],
         });
         payload_source.mark_unchanged();
 
@@ -106,7 +112,7 @@ impl SignatureAggregator {
         let aggregator = SignatureAggregator {
             id: config.id.clone(),
             implementation,
-            synthetics,
+            feeds,
             raft_client,
             signed_entries_source,
             payload_sink,
@@ -123,13 +129,13 @@ impl SignatureAggregator {
             }
         };
 
-        let payload_ages: DashMap<String, (SystemTime, SystemTime)> = self
-            .synthetics
+        let payload_ages: DashMap<String, (SystemTime, Option<SystemTime>)> = self
+            .feeds
             .iter()
-            .map(|synthetic| {
+            .map(|feed| {
                 let now = SystemTime::now();
-                let later = now + Duration::from_secs(60);
-                (synthetic.clone(), (now, later))
+                let later = Some(now + Duration::from_secs(60));
+                (feed.clone(), (now, later))
             })
             .collect();
         let payload_ages = Arc::new(payload_ages);
@@ -142,7 +148,7 @@ impl SignatureAggregator {
                     let (last_updated, valid_until) = *entry.value();
                     drop(entry);
 
-                    let is_valid: u64 = if SystemTime::now() < valid_until {
+                    let is_valid: u64 = if valid_until.is_none_or(|t| SystemTime::now() < t) {
                         1
                     } else {
                         0
@@ -163,63 +169,71 @@ impl SignatureAggregator {
         };
 
         let id = self.id;
-        let synthetics = self.synthetics;
+        let feeds = self.feeds;
         let raft_client = self.raft_client;
         let mut payload_source = self.signed_entries_source;
         let publish_task = async move {
-            // map of synthetic name to most recent payload produced for that entry
-            let mut latest_entries: BTreeMap<String, SyntheticPayloadEntry> = BTreeMap::new();
+            // map of feed to most recent entry produced for that feed
+            let mut latest_entries: BTreeMap<String, PayloadEntry> = BTreeMap::new();
 
-            // map of synthetic name to number of times it's failed in a row
-            let mut synthetic_failures: BTreeMap<String, usize> =
-                synthetics.iter().map(|s| (s.clone(), 0)).collect();
+            // map of feed name to number of times it's failed in a row
+            let mut all_failures: BTreeMap<String, usize> =
+                feeds.iter().map(|s| (s.clone(), 0)).collect();
 
             while let Some((publisher, payload)) = payload_source.recv().await {
                 let span = info_span!("publish_entries");
                 let mut updated = BTreeSet::new();
                 // update the payload age monitor now that we have a new payload
-                for entry in &payload.synthetics {
-                    let synthetic = entry.feed.data.synthetic.clone();
-                    let last_updated = entry.timestamp.unwrap_or(payload.timestamp);
+                let synthetics = payload.synthetics.into_iter().map(|s| {
+                    PayloadEntry::Synthetic(SyntheticPayloadEntry {
+                        timestamp: s.timestamp.unwrap_or(payload.timestamp),
+                        synthetic: s.feed.data.synthetic.clone(),
+                        price: s.price,
+                        payload: s.feed,
+                    })
+                });
+                let generics = payload.generics.into_iter().map(|i| {
+                    PayloadEntry::Generic(GenericPayloadEntry {
+                        timestamp: i.timestamp,
+                        name: i.feed.data.name.clone(),
+                        payload: i.feed,
+                    })
+                });
+                for entry in synthetics.chain(generics) {
+                    let feed = entry.feed();
                     if latest_entries
-                        .get(&synthetic)
-                        .is_some_and(|e| e.timestamp >= last_updated)
+                        .get(&feed)
+                        .is_some_and(|e| e.timestamp() >= entry.timestamp())
                     {
                         // This isn't actually a NEW payload...
                         continue;
                     }
-                    updated.insert(synthetic.clone());
-                    let price_feed_size = entry.feed.cbor_len(&mut ()) as u64;
-                    span.in_scope(|| {
-                        debug!(
-                            histogram.price_feed_size = price_feed_size,
-                            synthetic, "price feed size metrics"
-                        );
-                    });
-                    latest_entries.insert(
-                        synthetic.clone(),
-                        SyntheticPayloadEntry {
-                            timestamp: last_updated,
-                            synthetic: synthetic.clone(),
-                            price: entry.price,
-                            payload: entry.feed.clone(),
-                        },
-                    );
-                    let valid_until = find_end_of_entry_validity(entry);
-                    payload_ages.insert(synthetic, (last_updated, valid_until));
+                    updated.insert(feed.clone());
+                    if let PayloadEntry::Synthetic(synth) = &entry {
+                        let price_feed_size = synth.payload.cbor_len(&mut ()) as u64;
+                        span.in_scope(|| {
+                            debug!(
+                                histogram.price_feed_size = price_feed_size,
+                                synthetic = synth.synthetic,
+                                "price feed size metrics"
+                            );
+                        });
+                    }
+                    payload_ages.insert(feed.clone(), (entry.timestamp(), entry.valid_until()));
+                    latest_entries.insert(feed, entry);
                 }
 
                 if publisher == id {
                     // count the number of times we failed to publish some synthetic
                     let mut something_failed_too_often = false;
-                    for (synthetic, failures) in synthetic_failures.iter_mut() {
-                        if updated.contains(synthetic) {
+                    for (feed, failures) in all_failures.iter_mut() {
+                        if updated.contains(feed) {
                             *failures = 0;
                         } else {
                             *failures += 1;
                             span.in_scope(|| {
                                 warn!(
-                                    synthetic,
+                                    feed,
                                     failures, "Could not publish new price feed for synthetic"
                                 );
                             });
@@ -236,17 +250,23 @@ impl SignatureAggregator {
                         raft_client.abdicate();
                     }
                 } else {
-                    for failures in synthetic_failures.values_mut() {
+                    for failures in all_failures.values_mut() {
                         *failures = 0;
                     }
                 }
 
-                let entries = latest_entries.values().cloned().collect();
-                let payload = Payload {
+                let mut payload = Payload {
                     publisher,
                     timestamp: payload.timestamp,
-                    synthetics: entries,
+                    synthetics: vec![],
+                    generics: vec![],
                 };
+                for entry in latest_entries.values().cloned() {
+                    match entry {
+                        PayloadEntry::Synthetic(entry) => payload.synthetics.push(entry),
+                        PayloadEntry::Generic(entry) => payload.generics.push(entry),
+                    }
+                }
                 self.payload_sink.send_replace(payload);
             }
         };
@@ -259,8 +279,36 @@ impl SignatureAggregator {
     }
 }
 
-fn find_end_of_entry_validity(entry: &SyntheticEntry) -> SystemTime {
-    let upper_bound = entry.feed.data.validity.upper_bound;
+#[derive(Clone)]
+enum PayloadEntry {
+    Synthetic(SyntheticPayloadEntry),
+    Generic(GenericPayloadEntry),
+}
+impl PayloadEntry {
+    fn feed(&self) -> String {
+        match self {
+            Self::Synthetic(entry) => entry.synthetic.clone(),
+            Self::Generic(entry) => entry.name.clone(),
+        }
+    }
+
+    fn timestamp(&self) -> SystemTime {
+        match self {
+            Self::Synthetic(entry) => entry.timestamp,
+            Self::Generic(entry) => entry.timestamp,
+        }
+    }
+
+    fn valid_until(&self) -> Option<SystemTime> {
+        match self {
+            Self::Synthetic(entry) => Some(find_end_of_entry_validity(entry)),
+            Self::Generic(_) => None,
+        }
+    }
+}
+
+fn find_end_of_entry_validity(entry: &SyntheticPayloadEntry) -> SystemTime {
+    let upper_bound = entry.payload.data.validity.upper_bound;
     match upper_bound.bound_type {
         IntervalBoundType::NegativeInfinity => SystemTime::UNIX_EPOCH,
         IntervalBoundType::PositiveInfinity => SystemTime::now() + Duration::from_secs(60 * 525600),
@@ -280,8 +328,17 @@ pub struct SyntheticPayloadEntry {
 }
 
 #[derive(Serialize, Clone)]
+pub struct GenericPayloadEntry {
+    pub timestamp: SystemTime,
+    pub name: String,
+    #[serde(serialize_with = "cbor_encode_in_list")]
+    pub payload: Signed<GenericPriceFeed>,
+}
+
+#[derive(Serialize, Clone)]
 pub struct Payload {
     pub publisher: NodeId,
     pub timestamp: SystemTime,
+    pub generics: Vec<GenericPayloadEntry>,
     pub synthetics: Vec<SyntheticPayloadEntry>,
 }
